feat: include 'Raise hand' in call reactions menu
